applications/zpc/scripts/certification/requirements.txt
paho-mqtt
colorist 